requirements.txt
gspread
asyncpg
python-dotenv
PyYAML
google-auth
pydantic-settings
pydantic-settings
